requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
crewai==0.11.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
beautifulsoup4==4.12.2
aiohttp==3.9.1
asyncio==3.4.3
python-multipart==0.0.6
sqlalchemy==2.0.23
pysqlite3==0.5.4
openai==1.3.7
tavily-python==0.2.8
duckduckgo-search==4.1.1
newspaper3k==0.2.8
transformers==4.36.0
torch==2.1.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.28.1 